+++ components/hord-cli/src/core/mod.rs
@@ -1,11 +1,8 @@
 pub mod pipeline;
 pub mod protocol;
 
-use chainhook_sdk::bitcoincore_rpc::bitcoin::hashes::hex::FromHex;
-use chainhook_sdk::bitcoincore_rpc::bitcoin::Script;
-use chainhook_sdk::bitcoincore_rpc_json::bitcoin::{Txid, Witness};
 use chainhook_sdk::types::{
-    BitcoinBlockData, BitcoinNetwork, OrdinalInscriptionCurseType, OrdinalInscriptionRevealData,
+    BitcoinBlockData, BitcoinNetwork, OrdinalInscriptionRevealData,
     OrdinalOperation,
 };
 use dashmap::DashMap;
@@ -25,7 +22,7 @@ use chainhook_sdk::{
 use crate::config::{Config, LogConfig};
 
 use chainhook_sdk::indexer::bitcoin::{
-    standardize_bitcoin_block, BitcoinBlockFullBreakdown, BitcoinTransactionFullBreakdown,
+    standardize_bitcoin_block, BitcoinBlockFullBreakdown,
 };
 
 use crate::db::{
@@ -37,9 +34,8 @@ use crate::db::{
     insert_transfer_in_locations, remove_entry_from_blocks, remove_entry_from_inscriptions,
     LazyBlockTransaction,
 };
-use crate::ord::inscription_id::InscriptionId;
 
-use self::protocol::inscribing::InscriptionParser;
+use self::protocol::inscribing::parse_ordinal_operations;
 
 #[derive(Clone, Debug)]
 pub struct HordConfig {
@@ -73,86 +69,6 @@ pub fn parse_ordinals_and_standardize_block(
         }
     }
     Ok(block)
-}
-
-pub fn parse_ordinal_operations(
-    tx: &BitcoinTransactionFullBreakdown,
-    _ctx: &Context,
-) -> Vec<OrdinalOperation> {
-    // This should eventually become a loop once/if there is settlement on https://github.com/casey/ord/issues/2000.
-    let mut operations = vec![];
-    for (input_index, input) in tx.vin.iter().enumerate() {
-        if let Some(ref witness_data) = input.txinwitness {
-            let witness_data_hex: Vec<Vec<u8>> = witness_data
-                .iter()
-                .map(|w| hex::decode(w).unwrap())
-                .collect();
-            let witness = Witness::from_vec(witness_data_hex.clone());
-            let mut inscription = match InscriptionParser::parse(&witness) {
-                Ok(inscription) => inscription,
-                Err(_e) => {
-                    let mut cursed_inscription = None;
-                    for bytes in witness_data_hex.iter() {
-                        let script = Script::from(bytes.to_vec());
-                        let parser = InscriptionParser {
-                            instructions: script.instructions().peekable(),
-                        };
-
-                        let mut inscription = match parser.parse_script() {
-                            Ok(inscription) => inscription,
-                            Err(_) => continue,
-                        };
-                        inscription.curse = Some(OrdinalInscriptionCurseType::P2wsh);
-                        cursed_inscription = Some(inscription);
-                        break;
-                    }
-                    match cursed_inscription {
-                        Some(inscription) => inscription,
-                        None => continue,
-                    }
-                }
-            };
-
-            let inscription_id = InscriptionId {
-                txid: Txid::from_hex(&tx.txid).unwrap(),
-                index: input_index as u32,
-            };
-
-            if input_index > 0 {
-                inscription.curse = Some(OrdinalInscriptionCurseType::Batch);
-            }
-
-            let no_content_bytes = vec![];
-            let inscription_content_bytes = inscription.body().take().unwrap_or(&no_content_bytes);
-            let mut content_bytes = "0x".to_string();
-            content_bytes.push_str(&hex::encode(&inscription_content_bytes));
-
-            let payload = OrdinalInscriptionRevealData {
-                content_type: inscription.content_type().unwrap_or("unknown").to_string(),
-                content_bytes,
-                content_length: inscription_content_bytes.len(),
-                inscription_id: inscription_id.to_string(),
-                inscription_input_index: input_index,
-                tx_index: 0,
-                inscription_output_value: 0,
-                inscription_fee: 0,
-                inscription_number: 0,
-                inscriber_address: None,
-                ordinal_number: 0,
-                ordinal_block_height: 0,
-                ordinal_offset: 0,
-                transfers_pre_inscription: 0,
-                satpoint_post_inscription: format!(""),
-                curse_type: inscription.curse.take(),
-            };
-
-            operations.push(match &payload.curse_type {
-                Some(_) => OrdinalOperation::CursedInscriptionRevealed(payload),
-                None => OrdinalOperation::InscriptionRevealed(payload),
-            });
-        }
-    }
-    operations
 }
 
 pub fn get_inscriptions_revealed_in_block(
@@ -376,21 +292,4 @@ fn test_identify_next_output_index_destination() {
         ),
         SatPosition::Fee(2474)
     );
-}
-
-#[test]
-fn test_ordinal_inscription_parsing() {
-    let bytes = hex::decode("208737bc46923c3e64c7e6768c0346879468bf3aba795a5f5f56efca288f50ed2aac0063036f7264010118746578742f706c61696e3b636861727365743d7574662d38004c9948656c6c6f2030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030303030300a68").unwrap();
-
-    let script = Script::from(bytes);
-    let parser = InscriptionParser {
-        instructions: script.instructions().peekable(),
-    };
-
-    let inscription = match parser.parse_script() {
-        Ok(inscription) => inscription,
-        Err(_) => panic!(),
-    };
-
-    println!("{:?}", inscription);
 }
