+++ components/hord-cli/src/scan/bitcoin.rs
@@ -1,4 +1,4 @@
-use crate::archive::download_ordinals_dataset_if_required;
+use crate::download::download_ordinals_dataset_if_required;
 use crate::config::{Config, PredicatesApi};
 use crate::core::protocol::sequencing::{
     update_storage_and_augment_bitcoin_block_with_inscription_reveal_data_tx,
